+++ src/player-util.h
@@ -125,6 +125,8 @@ void disturb(struct player *p);
 void search(struct player *p);
 bool player_has_monster_in_view(const struct player *p);
 void player_forget_the_world(struct player *p);
+int player_power_chance(struct player *p, const struct player_power *power);
+bool player_use_power(struct player *p, struct player_power *power, int dir);
 void player_night_dream(struct player *p);
 void player_dream_chances(int law, int *bright, int *dark);