+++ src/ui-map.h
@@ -25,5 +25,6 @@ extern void display_map(int *cy, int *cx);
 extern void do_cmd_view_map(void);
 void do_cmd_view_world_map(void);
 void do_cmd_quest_log(void);
+void do_cmd_racial_power(void);
 void ui_enter_service(game_event_type type, game_event_data *data,
 					  void *user);